+++ tests/test_level4/test_indeterminate.py
@@ -20,22 +20,22 @@ class TestIndeterminate(util.TestCase):
         </div>
         <div>
           <input type="radio" name="test" id="radio1">
-          <label for="radio">This label starts out lime.</label>
+          <label for="radio1">This label starts out lime.</label>
           <form>
             <input type="radio" name="test" id="radio2">
-            <label for="radio">This label starts out lime.</label>
+            <label for="radio2">This label starts out lime.</label>
 
             <input type="radio" name="test" id="radio3" checked>
-            <label for="radio">This label starts out lime.</label>
+            <label for="radio3">This label starts out lime.</label>
 
             <input type="radio" name="other" id="radio4">
-            <label for="radio">This label starts out lime.</label>
+            <label for="radio4">This label starts out lime.</label>
 
             <input type="radio" name="other" id="radio5">
-            <label for="radio">This label starts out lime.</label>
+            <label for="radio5">This label starts out lime.</label>
           </form>
           <input type="radio" name="test" id="radio6">
-          <label for="radio">This label starts out lime.</label>
+          <label for="radio6">This label starts out lime.</label>
         </div>
         """
 
@@ -44,6 +44,33 @@ class TestIndeterminate(util.TestCase):
             ":indeterminate",
             ['checkbox', 'radio1', 'radio6', 'radio4', 'radio5', 'radio-no-name1'],
             flags=util.HTML
+        )
+
+    def test_iframe(self):
+        """Test indeterminate when `iframe` is involved."""
+
+        markup = """
+        <form>
+            <input type="radio" name="test" id="radio1">
+            <label for="radio1">This label starts out lime.</label>
+
+            <iframe>
+            <html>
+            <body>
+            <input type="radio" name="test" id="radio2" checked>
+            <label for="radio2">This label starts out lime.</label>
+
+            <input type="radio" name="other" id="radio3">
+            <label for="radio3">This label starts out lime.</label>
+            </body>
+            </html>
+            </iframe></form>"""
+
+        self.assert_selector(
+            markup,
+            ":indeterminate",
+            ['radio1', 'radio3'],
+            flags=util.PYHTML
         )