+++ docs/src/markdown/about/changelog.md
@@ -2,7 +2,7 @@
 
 ## 1.9.1
 
-- **FIX**: `:root`, `:default`, `:intermediate`, `:lang()`, and `:dir()` will properly account for HTML `iframe`
+- **FIX**: `:root`, `:default`, `:indeterminate`, `:lang()`, and `:dir()` will properly account for HTML `iframe`
 elements in their logic when selecting or matching an element. Their logic will be restricted to the document for which
 the element under consideration applies.
 - **FIX**: HTML pseudo-classes will check that all key elements checked are in the XHTML namespace (HTML parsers that do